refactor: Move PostProcessor to a dedicated file
